Major Tiny Heist cleanup
Commit message! Recently I've been working on a collection of some of my old games, which will include Tiny Heist. This game was such a complete mess, though, that it wasn't really feasible to port it as it was.

So, I decided to do a major clean up of it before I attempted a port. I've stripped lots and lots of unused code, and cleaned up as much of the rest as I could. Some highlights:
 - All the old "terrylib" stuff (a proto version of haxegon) has been removed or refactored
 - Old generator code that wasn't used in the final version has been stripped out
 - Ditto some old entity code for things that didn't make the final version of the game - Probes, Tripwires
 - I moved all the cached roomdata to a json file that's loaded when the game starts, (also the perlin noise data file, which is almost half a megabyte)
 - Lots of big classes have been split up into smaller ones to make it all easier to read - e.g. PopUp, Glitch, four separate states each game state
 - The rest, I moved into folders so that it's at least all in one place - world/ and entities/ rather than split across two or three different places

This is a substantially easier to understand project now, and hopefully a better example project for anyone out there (?) interested in using Haxegon.